+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/realtime/RealtimeControllerFacade.kt
@@ -82,4 +82,21 @@ interface RealtimeControllerFacade {
      * @param topic: String - topic of messages for unsubscription
      */
     fun removeRealtimeDataMessageObserverFromTopic(topic: String)
+
+    /**
+     * Enable or disable Voice Focus (ML-based noise suppression) on the audio input.
+     *
+     * Voice Focus is disabled by default when audioClient starts.
+     *
+     * @param enabled: Boolean - whether to enable Voice Focus or not
+     * @return Boolean whether the enable/disable action succeeded
+     */
+    fun realtimeSetVoiceFocusEnabled(enabled: Boolean): Boolean
+
+    /**
+     * Checks if Voice Focus is enabled.
+     *
+     * @return Boolean whether Voice Focus is enabled or not
+     */
+    fun realtimeIsVoiceFocusEnabled(): Boolean
 }