+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/model/MeetingModel.kt
@@ -25,6 +25,7 @@ class MeetingModel : ViewModel() {
     var isMuted = false
     var isCameraOn = false
     var isDeviceListDialogOn = false
+    var isAdditionalOptionsDialogOn = false
     var lastReceivedMessageTimestamp = 0L
     var tabIndex = 0
 }