+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-### Unreleased
+## Unreleased
+
+### Added
+* Added new APIs in `RealtimeControllerFacade` to enable/disable Voice Focus (ML-based noise suppression) and get the on/off status of Voice Focus
+* Added Voice Focus feature in Android demo app
 
 ### Fixed
 * **Breaking** Changed behavior to no longer call `onVideoTileSizeChanged` when a video is paused to fix a bug where pausing triggered this callback with width=0 and height=0